+++ model/jpa/src/main/java/org/keycloak/models/jpa/RealmAdapter.java
@@ -1060,7 +1060,7 @@ public class RealmAdapter implements LegacyRealmModel, JpaModel<RealmEntity> {
 
     @Override
     public Stream<OpenIdFederationConfig> getTrustAnchorsBasedOnTypes(EntityTypeEnum entityType, ClientRegistrationTypeEnum clientRegistrationType) {
-        return isOpenIdFederationEnabled() ? getOpenIdFederations().stream().filter(x -> x.getEntityTypes().contains(entityType) && x.getClientRegistrationTypesSupported().contains(clientRegistrationType)) : Stream.empty();
+        return isOpenIdFederationEnabled() ? getOpenIdFederations().stream().filter(x -> entityType.equals(x.getEntityType()) && x.getClientRegistrationTypesSupported().contains(clientRegistrationType)) : Stream.empty();
     }
 
     @Override
@@ -1134,12 +1134,10 @@ public class RealmAdapter implements LegacyRealmModel, JpaModel<RealmEntity> {
             OpenIdFederationConfig fedConfig = new OpenIdFederationConfig();
             fedConfig.setInternalId(fedEntity.getInternalId());
             fedConfig.setTrustAnchor(fedEntity.getTrustAnchor());
+            fedConfig.setEntityType(fedEntity.getEntityType());
             String clientRegistrationTypesSupportedStr = fedEntity.getConfig().get(Constants.OPENID_FEDERATION_CLIENT_REGISTRATION_TYPES_SUPPORTED);
             List<String> clientRegistrationTypesSupported = (clientRegistrationTypesSupportedStr == null || clientRegistrationTypesSupportedStr.isEmpty()) ? new ArrayList<>() : Arrays.asList(clientRegistrationTypesSupportedStr.split("##"));
             fedConfig.setClientRegistrationTypesSupported(clientRegistrationTypesSupported.stream().map(x -> ClientRegistrationTypeEnum.valueOf(x)).collect(Collectors.toList()));
-            String entityTypesStr = fedEntity.getConfig().get(Constants.OPENID_FEDERATION_ENTITY_TYPES);
-            List<String> entityTypes = (entityTypesStr == null || entityTypesStr.isEmpty()) ? new ArrayList<>() : Arrays.asList(entityTypesStr.split("##"));
-            fedConfig.setEntityTypes(entityTypes.stream().map(x -> EntityTypeEnum.valueOf(x)).collect(Collectors.toList()));
             fedConfig.setIdpConfiguration(fedEntity.getIdpConfiguration());
             return fedConfig;
         }).collect(Collectors.toList());
@@ -1164,7 +1162,7 @@ public class RealmAdapter implements LegacyRealmModel, JpaModel<RealmEntity> {
 
     private void transformToOpenIdFederationEntity(OpenIdFederationEntity fedEntity, OpenIdFederationConfig fedConfig){
         fedEntity.setTrustAnchor(fedConfig.getTrustAnchor());
-        fedEntity.getConfig().put(Constants.OPENID_FEDERATION_ENTITY_TYPES, fedConfig.getEntityTypes().stream().map(x -> x.name()).collect(Collectors.joining("##")));
+        fedEntity.setEntityType(fedConfig.getEntityType());
         fedEntity.getConfig().put(Constants.OPENID_FEDERATION_CLIENT_REGISTRATION_TYPES_SUPPORTED, fedConfig.getClientRegistrationTypesSupported().stream().map(x -> x.name()).collect(Collectors.joining("##")));
         fedEntity.setIdpConfiguration(fedConfig.getIdpConfiguration());
         fedEntity.setRealm(realm);